UDC_Data-Mining/requirements.txt
py-cord == 2.2.2
discord.py == 2.4.0
requests == 2.32.3